made everything seperate for a better look.
