+++ src/components/context/ThemeContext.tsx
@@ -1,0 +1,26 @@
+// src/components/context/ThemeContext.tsx
+import React, { createContext, useContext } from "react";
+
+interface ThemeContextType {
+  theme: string;
+  setTheme: (theme: string) => void;
+}
+
+const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
+
+export const ThemeProvider: React.FC<{
+  children: React.ReactNode;
+  value: ThemeContextType;
+}> = ({ children, value }) => {
+  return (
+    <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>
+  );
+};
+
+export const useTheme = () => {
+  const context = useContext(ThemeContext);
+  if (!context) {
+    throw new Error("useTheme must be used within a ThemeProvider");
+  }
+  return context;
+};
